Move and simplify test_scan
