+++ python/cudf/cudf/tests/series/methods/test_cummulative_methods.py
@@ -1,0 +1,124 @@
+# Copyright (c) 2025, NVIDIA CORPORATION.
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import cudf
+from cudf.core.dtypes import Decimal32Dtype, Decimal64Dtype, Decimal128Dtype
+from cudf.testing import assert_eq
+from cudf.testing._utils import gen_rand
+
+
+@pytest.fixture(params=[0, 5])
+def nelem(request):
+    return request.param
+
+
+@pytest.fixture(params=["cumsum", "cummin", "cummax", "cumprod"])
+def cummulative_methods(request):
+    return request.param
+
+
+def test_cummulative_methods(numeric_types_as_str, nelem, cummulative_methods):
+    dtype = np.dtype(numeric_types_as_str)
+    if dtype == np.int8:
+        # to keep data in range
+        data = gen_rand(dtype, nelem, low=-2, high=2)
+    else:
+        data = gen_rand(dtype, nelem)
+
+    decimal = 4 if dtype == np.float32 else 6
+
+    gs = cudf.Series(data)
+    ps = pd.Series(data)
+    np.testing.assert_array_almost_equal(
+        getattr(gs, cummulative_methods)().to_numpy(),
+        getattr(ps, cummulative_methods)(),
+        decimal=decimal,
+    )
+
+
+def test_cummulative_methods_masked(numeric_types_as_str, cummulative_methods):
+    data = [1, 2, None, 4, 5]
+    gs = cudf.Series(data).astype(numeric_types_as_str)
+    # float64 since pandas usses NaN as missing value
+    ps = pd.Series(data).astype("float64")
+    assert_eq(
+        getattr(gs, cummulative_methods)(),
+        getattr(ps, cummulative_methods)(),
+        check_dtype=False,
+    )
+
+
+@pytest.mark.parametrize(
+    "dtype",
+    [
+        Decimal64Dtype(8, 4),
+        Decimal64Dtype(10, 5),
+        Decimal64Dtype(12, 7),
+        Decimal32Dtype(8, 5),
+        Decimal128Dtype(13, 6),
+    ],
+)
+def test_cumsum_decimal(dtype):
+    data = ["243.32", "48.245", "-7234.298", np.nan, "-467.2"]
+    gser = cudf.Series(data).astype(dtype)
+    pser = pd.Series(data, dtype="float64")
+
+    got = gser.cumsum()
+    expected = cudf.Series.from_pandas(pser.cumsum()).astype(dtype)
+
+    assert_eq(got, expected)
+
+
+@pytest.mark.parametrize(
+    "dtype",
+    [
+        Decimal64Dtype(8, 4),
+        Decimal64Dtype(11, 6),
+        Decimal64Dtype(14, 7),
+        Decimal32Dtype(8, 4),
+        Decimal128Dtype(11, 6),
+    ],
+)
+def test_cummin_decimal(dtype):
+    data = ["8394.294", np.nan, "-9940.444", np.nan, "-23.928"]
+    gser = cudf.Series(data).astype(dtype)
+    pser = pd.Series(data, dtype="float64")
+
+    got = gser.cummin()
+    expected = cudf.Series.from_pandas(pser.cummin()).astype(dtype)
+
+    assert_eq(got, expected)
+
+
+@pytest.mark.parametrize(
+    "dtype",
+    [
+        Decimal64Dtype(8, 4),
+        Decimal64Dtype(11, 6),
+        Decimal64Dtype(14, 7),
+        Decimal32Dtype(8, 4),
+        Decimal128Dtype(11, 6),
+    ],
+)
+def test_cummax_decimal(dtype):
+    data = [np.nan, "54.203", "8.222", "644.32", "-562.272"]
+    gser = cudf.Series(data).astype(dtype)
+    pser = pd.Series(data, dtype="float64")
+
+    got = gser.cummax()
+    expected = cudf.Series.from_pandas(pser.cummax()).astype(dtype)
+
+    assert_eq(got, expected)
+
+
+@pytest.mark.parametrize("method", ["cumsum", "cumprod"])
+def test_scan_boolean(method):
+    s = cudf.Series([True, False, True, False])
+
+    got = getattr(s, method)()
+    expect = getattr(s.to_pandas(), method)()
+
+    assert_eq(expect, got)
